Add files via upload
I was STRUGGLING to make this even passable, I abhor this "bootstrap" stuff and I'm sorry but I'd rather just code it by hand in the future. I have so much less control this way and I can't seem to make anything behave like it normally would. Not my cup of tea.